requirements.txt
Flask>=3.0.3,<3.1
pillow>=11.0.0,<11.1